feat: add a wasm specific entry point, error handler and browser friendly builtins
