+++ object/error_wasm.go
@@ -1,14 +1,12 @@
-//go:build !js || !wasm
+//go:build wasm && js
 package object
-
-import "fmt"
 
 type Error struct {
 	Message string
 }
 
 func (e *Error) Inspect() string {
-	msg := fmt.Sprintf("\x1b[%dm%s\x1b[0m", 31, "Kosa: ")
-	return msg + e.Message
+	// msg := fmt.Sprintf("\x1b[%dm%s\x1b[0m", 31, "Kosa: ") // removes ANSI codes
+	return "Kosa: " + e.Message
 }
 func (e *Error) Type() ObjectType { return ERROR_OBJ }
